Ensure docker service is stopped or started
This change modifies restarting of docker and
bootstrap docker by adding checks to ensure
the service has stopped or started before
continuing the scripts. Restarting
bootstrap docker also includes a check on
systemctl to handle stopping and starting of
the service.

This is done to prevent a potential race condition
where the script continues to bring up Kubernetes
before the restart of either the bootstrap docker
or docker service is complete causing Kubernetes to
fail mid-operation due to docker being stopped.