Update kubeconfig and kubelet parameters for K8s 1.6.0 upgrade
This change modifies the kubeconfig which is used by the Kubelet to
communicate with API server to have username and password. This change
also modifies one of the command line parameters to start the Kubelet
running on a K8s master node. These changes are needed for our upgrade
to K8s 1.6.0.